JAMES-2586 Rename postgres-app tests' name: JPA -> Postgres
